PROGETTO FINALE SIGNORE-FERRAUTO/grpc_server/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
yfinance
prometheus-client
